+++ api/close_contact/handlers.js
@@ -1,9 +1,11 @@
 const { HTTP } = require('../../helpers/constant')
 const replyHelper = require('../helpers')
+const Helper = require('../../helpers/custom')
+
 module.exports = (server) => {
     return {
         /**
-         * GET /api/cases/{caseId}/close-contacts
+         * GET /api/close-contacts
          * @param {*} request
          * @param {*} reply
          */
@@ -30,7 +32,7 @@ module.exports = (server) => {
                 })
         },
         /**
-         * POST /api/cases/{id}/close-contacts
+         * POST /api/cases/{caseId}/close-contacts
          * @param {*} request
          * @param {*} reply
          */
@@ -41,39 +43,61 @@ module.exports = (server) => {
                 request.payload,
                 (err, result) => {
                     if (err) return replyHelper.errorResponse(reply, err)
-                    return replyHelper.successResponse(reply, result, HTTP.CREATED)
+
+                    server.methods.services.closeContactHistories.create(
+                        result._id,
+                        request.payload.latest_history,
+                        (err, resultChild) => {
+                            if (err) return replyHelper.errorResponse(reply, err)
+                            
+                            const res = Object.assign(result, { latest_history: resultChild })
+                            return replyHelper.successResponse(reply, res, HTTP.CREATED)
+                        })
                 })
         },
         /**
-         * POST /api/cases/{id}/close-contacts-with-report
+         * GET /api/cases/{caseId}/close-contacts/{closeContactId}
          * @param {*} request
          * @param {*} reply
          */
-        async CreateCloseContactWithReport(request, reply){
-            server.methods.services.closeContacts.create(
-                request.params.caseId,
+        async DetailCloseContact(request, reply) {
+            server.methods.services.closeContacts.show(
+                request.params.closeContactId,
+                (err, result) => {
+                    if (err) return replyHelper.errorResponse(reply, err)
+                    return replyHelper.successResponse(reply, result, HTTP.OK)
+                })
+        },
+        /**
+         * PUT /api/cases/{caseId}/close-contacts/{closeContactId}
+         * @param {*} request
+         * @param {*} reply
+         */
+        async UpdateCloseContact(request, reply){
+            const currentHistory = request.pre.close_contact.latest_history
+            const requestHistory = request.payload.latest_history
+            const isDirty = Helper.isDirty(currentHistory, requestHistory)
+            server.methods.services.closeContacts.update(
+                request.params.closeContactId,
                 request.auth.credentials.user,
                 request.payload,
-                (err, closeContact) => {
+                (err, result) => {
                     if (err) return replyHelper.errorResponse(reply, err)
 
-                    server.methods.services.closeContactReport.create(
-                        closeContact._id,
-                        request.auth.credentials.user,
-                        request.payload,
-                        (err, report) => {
-                            if (err) return replyHelper.errorResponse(reply, err)
-
-                            server.methods.services.closeContactReportHistories.create(
-                                report._id,
-                                request.payload.latest_report_history,
-                                (err, history) => {
-                                    if (err) return replyHelper.errorResponse(reply, err)
-
-                                    const res = Object.assign(report, { latest_report_history: history })
-                                    return replyHelper.successResponse(reply, res, HTTP.CREATED)
-                                })
-                        })
+                    if (!requestHistory || !isDirty) {
+                        const res = Object.assign(result, { latest_history: currentHistory })
+                        return replyHelper.successResponse(reply, res, HTTP.OK)
+                    } else {
+                        server.methods.services.closeContactHistories.create(
+                            result._id,
+                            requestHistory,
+                            (err, resultChild) => {
+                                if (err) return replyHelper.errorResponse(reply, err)
+                                
+                                const res = Object.assign(result, { latest_history: resultChild })
+                                return replyHelper.successResponse(reply, res, HTTP.OK)
+                            })
+                    }
                 })
         },
         /**
@@ -83,13 +107,12 @@ module.exports = (server) => {
          */
         async DeleteCloseContact(request, reply) {          
             server.methods.services.closeContacts.delete(
-                request.params.id,
+                request.params.closeContactId,
                 request.auth.credentials.user,
-                request.payload,
                 (err, result) => {
                     if (err) return replyHelper.errorResponse(reply, err)
                     return replyHelper.successResponse(reply, result, HTTP.OK)
                 })
-        },
+        }
     }
 }
