+++ api/close_contact/package.json
@@ -1,4 +1,4 @@
 {
-    "name" : "close_contact",
+    "name" : "close_contact_report",
     "version": "1.0.0"
 }
