+++ api/close_contact/route_prerequesites.js
@@ -20,6 +20,26 @@ const getCasebyId = server => {
     }
 }
 
+const getCloseContactbyId = server => {
+    return {
+        method: (request, reply) => {
+             let id = request.params.closeContactId
+             server.methods.services.closeContacts
+                .show(id, (err, result) => {
+                    if (err) return replyHelper.errorResponse(reply, err).takeover()
+                    if (!result) {
+                        return replyHelper.messageResponse(reply,
+                            ERRORS.INVALID.PARAMS_VALUE, HTTP.UNPROCESSABLE_ENTITY
+                        ).takeover()
+                    }
+                    return reply(result)
+                })
+        },
+        assign: 'close_contact'
+    }
+}
+
 module.exports = {
-    getCasebyId
+    getCasebyId,
+    getCloseContactbyId
 }
