+++ api/close_contact/routes.js
@@ -1,9 +1,9 @@
 const inputValidations = require('./validations/input')
-const reportInputValidations = require('../close_contact_report/validations/input')
 
 module.exports = (server) =>{
     const handlers = require('./handlers')(server)
-    const getCasebyId = require('./route_prerequesites').getCasebyId(server)
+    const getCaseById = require('./route_prerequesites').getCasebyId(server)
+    const getCloseContactbyId = require('./route_prerequesites').getCloseContactbyId(server)
 
     return [
         {
@@ -12,7 +12,8 @@ module.exports = (server) =>{
             config: {
                 auth: 'jwt',
                 description: 'show list of all close-contacts',
-                tags: ['api', 'cases', 'close.contacts']
+                tags: ['api', 'cases', 'close.contacts'],
+                validate: inputValidations.QueryValidations
             },
             handler: handlers.ListCloseContact
         },
@@ -27,39 +28,58 @@ module.exports = (server) =>{
             handler: handlers.ListCloseContactCase
         },
         {
+            method: 'GET',
+            path: '/cases/{caseId}/close-contacts/{closeContactId}',
+            config: {
+                auth: 'jwt',
+                description: 'show a specific close contact',
+                tags: ['api', 'cases', 'close.contacts'],
+                pre: [
+                    getCaseById
+                ]
+            },
+            handler: handlers.DetailCloseContact
+        },
+        {
             method: 'POST',
             path: '/cases/{caseId}/close-contacts',
+            config: {
+                auth: 'jwt',
+                description: 'create new close contacts report',
+                tags: ['api', 'cases', 'close.contacts.reports'],
+                validate: inputValidations.RequestPayload,
+                pre: [ getCaseById ]
+            },
+            handler: handlers.CreateCloseContact
+        },
+        {
+            method: 'PUT',
+            path: '/cases/{caseId}/close-contacts/{closeContactId}',
             config: {
                 auth: 'jwt',
                 description: 'create new close contacts',
                 tags: ['api', 'cases', 'close.contacts'],
                 validate: inputValidations.RequestPayload,
-                pre: [ getCasebyId ]
+                pre: [
+                    getCaseById,
+                    getCloseContactbyId
+                ]
             },
-            handler: handlers.CreateCloseContact
-        },
-        {
-            method: 'POST',
-            path: '/cases/{caseId}/close-contacts-with-report',            
-            config: {
-                auth: 'jwt',
-                description: 'create new close contacts',
-                tags: ['api', 'cases', 'close.contacts'],
-                validate: reportInputValidations.RequestPayload,
-                pre: [ getCasebyId ]
-            },
-            handler: handlers.CreateCloseContactWithReport
+            handler: handlers.UpdateCloseContact
         },
         {
             method: 'DELETE',
-            path: '/cases/{caseId}/close-contacts/{id}',
+            path: '/cases/{caseId}/close-contacts/{closeContactId}',
             config: {
                 auth: 'jwt',
-                description: 'show a specific cases history',
+                description: 'show a specific close contact',
                 tags: ['api', 'cases', 'close.contacts'],
-                pre: [ getCasebyId ]
+                pre: [
+                    getCaseById,
+                    getCloseContactbyId
+                ]
             },
             handler: handlers.DeleteCloseContact
-        }
+        },
     ]
 }
