--- api/close_contact_report/index.js
@@ -1,12 +1,0 @@
-const Routes = require('./routes')
-
-const register = (server, options, next) =>{
-    server.route(Routes(server))
-    return next()
-}
-
-register.attributes ={
-    pkg: require('./package.json')
-}
-
-module.exports = register
