--- api/close_contact_report/package.json
@@ -1,4 +1,0 @@
-{
-    "name" : "close_contact_report",
-    "version": "1.0.0"
-}
